MSVC warning mitigation.
Fixes #120.
Fixes #121.
Fixes #122.
Fixes #123.

+++ include/boost/parser/detail/printing_impl.hpp
@@ -226,7 +226,7 @@ namespace boost { namespace parser { namespace detail {
                 detail::print_parser(context, parser, os, components);
                 ++components;
                 ++i;
-                prev_group = group;
+                prev_group = (int)group;
             });
         if (prev_group && !printed_ellipsis)
             os << ']';
